+++ lizard_connector/connector.py
@@ -14,6 +14,10 @@ Includes:
 
 import functools
 import json
+
+import time
+from threading import Thread, RLock
+
 try:
     from urllib.parse import urlencode
     from urllib.parse import urljoin
@@ -27,13 +31,38 @@ except ImportError:
 
 import lizard_connector.queries
 
+ASYNC_POLL_TIME = 1
+ASYNC_POLL_TIME_INCREASE = 1.5
+
 
 class LizardApiTooManyResults(Exception):
     pass
 
 
+class LizardApiAsyncTaskFailure(Exception):
+    pass
+
+
 class InvalidUrlError(Exception):
     pass
+
+
+def no_op(*args, **kwargs):
+    pass
+
+
+def save_to_json(result, file_base="api_result"):
+    """
+    Saves a result to json with a timestamp in milliseconds.
+
+    Args:
+        result (list|dict): a json dumpable object to save to file.
+        file_base (str): filename base. Can contain a relative or absolute
+                         path.
+    """
+    filename = "{}_{}.json".format(file_base, str(int(time.time() * 1000)))
+    with open(filename, 'w') as json_filehandler:
+        json.dump(result, json_filehandler)
 
 
 class Connector(object):
@@ -114,6 +143,7 @@ class Connector(object):
             request_obj = urllib_request.Request(url, headers=self.header)
         resp = urlopen(request_obj)
         content = resp.read().decode('UTF-8')
+        # print(content)
         return json.loads(content)
 
     @property
@@ -150,9 +180,9 @@ class PaginatedRequest(object):
                             is used.
             paginate (bool): when set to True, paginate results. The connector
                              now can be used as an iterator also the next_page
-                             method can be used to obtain the next page. When 
-                             set to False a query is built with a max size of 
-                             max_results. All these results are returned at 
+                             method can be used to obtain the next page. When
+                             set to False a query is built with a max size of
+                             max_results. All these results are returned at
                              once.
         """
         self._endpoint = endpoint
@@ -213,9 +243,9 @@ class Endpoint(Connector):
                             is used.
             paginate (bool): when set to True, paginate results. The connector
                              now can be used as an iterator also the next_page
-                             method can be used to obtain the next page. When 
-                             set to False a query is built with a max size of 
-                             max_results. All these results are returned at 
+                             method can be used to obtain the next page. When
+                             set to False a query is built with a max size of
+                             max_results. All these results are returned at
                              once.
         """
         super(Endpoint, self).__init__(**kwargs)
@@ -231,7 +261,7 @@ class Endpoint(Connector):
             page_size=self.max_results)
         q.update(*querydicts, **queries)
         query = "?" + urlencode(q)
-        return  urljoin(self.base_url, query)
+        return urljoin(self.base_url, query)
 
     def download(self, *querydicts, **queries):
         """
@@ -248,9 +278,106 @@ class Endpoint(Connector):
         url = self._build_url(*querydicts, **queries)
         return self.get(url)
 
-    def download_paginated(self, page_size=10, *querydicts, **queries):
+    def download_paginated(self, page_size=100, *querydicts, **queries):
+        """
+        Instantiates an iterable paginated request.
+
+        The iterable returned has a length after first use. Example::
+
+            end_point = Endpoint("sluices")
+            paginated_request = end_point.download_paginated()
+            len(paginated_request)  # this is None
+            first_page = next(paginated_request)
+            # this results in a number (the amount of available sluices):
+            len(paginated_request)
+
+        Args:
+            querydicts (iterable): all key valuepairs from dictionaries are
+                                   used as queries.
+            page_size (int): number of results per iteration.
+            lock (Lock): a threading lock
+            thread_class (class): either Thread or a similar Thread class like
+                                  QThread
+            queries (dict): all keyword arguments are used as queries.
+        Returns:
+            an iterable that returns the results of each page.
+        """
         url = self._build_url(page_size=page_size, *querydicts, **queries)
         return PaginatedRequest(self, url)
+
+    def download_async(self, call_back=None, lock=None, *querydicts,
+                       **queries):
+        """
+        Downloads async via a Thread. A call_back function handles the results.
+
+        By default download_async does make a call, but doesn't do anything. We
+        provide a default method to save to file: save_to_json.
+
+        Args:
+            querydicts (iterable): all key valuepairs from dictionaries are
+                                   used as queries.
+            call_back (function): call back function that is called with the
+                                  downloaded result
+            lock (Lock): a threading lock. This lock is used when executing the
+                         call back function.
+            queries (dict): all keyword arguments are used as queries.
+        """
+        if call_back is None:
+            call_back = no_op
+        if lock is None:
+            lock = RLock()
+        args = (call_back, lock) + querydicts
+        thread = Thread(
+            target=self.async_worker,
+            args=args,
+            kwargs=queries
+        )
+        thread.start()
+
+    def _poll_task(self, task_url):
+        poll_result = self.get(task_url)
+        task_status = poll_result.get("task_status")
+        sleep_time = ASYNC_POLL_TIME
+        if task_status == "PENDING":
+            time.sleep(sleep_time)
+            sleep_time *= ASYNC_POLL_TIME_INCREASE
+            return None, True
+        elif task_status == "SUCCESS":
+            url = poll_result.get('result_url')
+            return self.get(url), False
+        raise LizardApiAsyncTaskFailure(task_status, task_url)
+
+    def async_worker(self, call_back, lock=None, *querydicts, **queries):
+        """
+        Starts a download as an api async task, but handles it synchronously.
+        A call_back function handles the results.
+
+        This function does not use threading and is provided to use in other
+        threads or processes other than the Thread from the python standard
+        library like QThread.
+
+        By default async_worker does make a call, but doesn't do anything. We
+        provide a default method to save to file: save_to_json.
+
+        Args:
+            querydicts (iterable): all key valuepairs from dictionaries are
+                                   used as queries.
+            call_back (function): call back function that is called with the
+                                  downloaded result
+            lock (Lock): a threading lock. This lock is used when executing the
+                         call back function.
+            queries (dict): all keyword arguments are used as queries.
+        """
+        queries.update({"async": "true"})
+        task_url = self.download(*querydicts, **queries).get('url')
+        keep_polling = True
+        while keep_polling:
+            result, keep_polling = self._poll_task(task_url)
+        if lock:
+            with lock:
+                call_back(result)
+        else:
+            call_back(result)
 
     def upload(self, data, uuid=None):
         """